Fully remove old useCcreData
